🔨 chore(Minecraft_Server_Launcher_2): update compiled class files
🔥 chore: remove unnecessary binary files in production folder

🐛 fix(WorldCopyHandler.class): update binary file with changes

🐛 fix(AddWorldsPanel.java): replace exStackTraceToString with getErrorDialogMessage in alert calls
✨ feat(AddWorldsPanel.java): add serverNameAndStuff JPanel to display server world name and stuff

🐛 fix(AddWorldsPanel.java): replace exStackTraceToString with getErrorDialogMessage to show error message in alert dialog

🔨 refactor(AddWorldsPanel.java): refactor GUI layout to improve readability and spacing

🐛 fix(AddWorldsPanel.java): remove unused paintComponent method and fix indentation
✨ feat(AddWorldsPanel.java): add setIcons method to set icons for selected and server worlds

🔧 chore(AddWorldsPanel.java): comment out unused code and fix issue with worldPanel border

🐛 fix(ButtonPanel.java): fix addWorldsPanel reference error
✨ feat(ButtonPanel.java): add server launch on button click

🔨 refactor(Frame.java): rename exStackTraceToString method to getErrorDialogMessage and add Exception parameter
✨ feat(Frame.java): add ActionListener to openServer button to debug LevelNameColorConverter

🐛 fix(Frame.java): replace exStackTraceToString with getErrorDialogMessage method in alert messages

🐛 fix(Frame.java): add InterruptedException to main method signature
🚀 chore(Frame.java): comment out unused code and remove unused variable
✨ feat(Frame.java): instantiate Frame class in main method

🐛 fix(ServerSelectionPanel.java): remove unused import statement
🚀 chore(ServerSelectionPanel.java): add public access modifier to addWorldsPanel variable
🐛 fix(ServerSelectionPanel.java): replace exStackTraceToString method with getErrorDialogMessage method in catch block
🚀 chore(ServerSelectionPanel.java): call setIcons method on addWorldsPanel after updating selectedIndexInComboBox variable

🎨 style(ServerSelectionPanel.java): add whitespace for readability
✨ feat(ServerSelectionPanel.java): add setIcons method to AddWorldsPanel in setPanels method

🐛 fix(ServerSelector.java): replace exStackTraceToString with getErrorDialogMessage in alert message

🔨 refactor(LevelNameColorConverter.java): remove unused import statements
🚀 chore(LevelNameColorConverter.java): make LevelNameColorConverter class final

🐛 fix(LevelNameColorConverter.java): fix reset code to properly close all open tags
✨ feat(LevelNameColorConverter.java): add method to convert Minecraft color codes to HTML tags

🔥 refactor(NBTParser.java): remove unused imports and methods
👌 style(NBTParser.java): reformat code with consistent indentation and spacing

🐛 fix(NBTParser.java): handle case when level.dat file is not found and update serverLevelName to reflect that

🔧 refactor(ServerPropertiesFile.java): change import of getErrorDialogMessage method from Gui.Frame and remove unused imports

🐛 fix(Runner.java): rename exStackTraceToString method to getErrorDialogMessage in import and usage

🐛 fix(WorldCopyHandler.java): replace exStackTraceToString with getErrorDialogMessage in alert messages

🐛 fix(WorldCopyHandler.java): replace exStackTraceToString with getErrorDialogMessage to display error message in alerts

🎉 feat(world_temp): add level_1.18.1.dat file to world_temp directory

+++ src/SelectedServer/NBTParser.java
@@ -1,16 +1,16 @@
 package SelectedServer;
 
 import Gui.AlertType;
-import Gui.Frame;
 import dev.dewy.nbt.Nbt;
 import dev.dewy.nbt.tags.collection.CompoundTag;
 import org.apache.commons.io.FileUtils;
 
+import javax.swing.*;
 import java.io.File;
 import java.io.IOException;
 
 import static Gui.Frame.alert;
-import static Gui.Frame.exStackTraceToString;
+import static Gui.Frame.getErrorDialogMessage;
 
 public class NBTParser extends Thread {
     private String levelName;
@@ -22,35 +22,34 @@ public class NBTParser extends Thread {
     @Override
     public void run() {
         System.out.println("Level.dat file location: " + ServerDetails.serverLevelDatFile);
-        System.out.println("Server Path: " + ServerDetails.serverPath);
-        System.out.println("Server World Path: " + ServerDetails.serverWorldPath);
-        System.out.println("Selected Server: " + ServerDetails.serverName);
         File pathToCopiedLevelDat = new File("world_temp\\level_" + ServerDetails.serverName + ".dat");
+        System.out.println("Path to copied level.dat: " + pathToCopiedLevelDat);
         try {
-            FileUtils.copyFile(new File(ServerDetails.serverLevelDatFile), pathToCopiedLevelDat);
-        } catch (IOException e) {
-            alert(AlertType.ERROR, "Cannot copy level.dat file." + exStackTraceToString(e.getStackTrace()));
+            if(new File(ServerDetails.serverLevelDatFile).exists())
+                FileUtils.copyFile(new File(ServerDetails.serverLevelDatFile), pathToCopiedLevelDat);
+            else
+                ServerDetails.serverLevelName = "Level.dat file not found";
+        } catch (Exception e) {
+            SwingUtilities.invokeLater(() -> alert(AlertType.ERROR, "Cannot copy level.dat file." + getErrorDialogMessage(e))); //issue #61 fix, it was conflicting with ui dispatch thread
         }
         Nbt levelDat = new Nbt();
         CompoundTag layerOne = null;
         try {
-            System.out.println(new File(ServerDetails.serverLevelDatFile));
-            System.out.println(pathToCopiedLevelDat);
-            layerOne = levelDat.fromFile(pathToCopiedLevelDat);
+            if(pathToCopiedLevelDat.exists())
+                layerOne = levelDat.fromFile(pathToCopiedLevelDat);
         } catch (IOException e) {
-//            if(Frame.isFrameInitialized)
-//                alert(AlertType.ERROR, exStackTraceToString(e.getStackTrace()));
-//            return;
+            SwingUtilities.invokeLater(() -> alert(AlertType.ERROR, getErrorDialogMessage(e))); //issue #61 fix, it was conflicting with ui dispatch thread
+            return;
         }
         if(layerOne != null) {
             CompoundTag levelDatContent = layerOne.get("Data");
             this.levelName = String.valueOf(levelDatContent.get("LevelName")).split("\"")[1];
         }
         System.out.println("Level name: " + levelName);
-        System.out.println();
+        System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
     }
 
     public String getLevelName() {
-        return levelName;
+        return LevelNameColorConverter.convertColors(levelName);
     }
 }
